% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractW}
\alias{extractW}
\title{Extracts W Columns from Matrix of Observations}
\usage{
extractW(mat)
}
\arguments{
\item{mat}{A \code{matrix} of observations, as the \code{obs} argument of
\code{function} \code{\link{tmle.npvi}}.}
}
\value{
The \code{matrix} extracted from \code{mat} by removing the two
\code{X} and \code{Y} columns.
}
\description{
Extracts the \code{W} column(s) from a \code{matrix} of observations.
}
\details{
Mainly for internal use.
}
\author{
Antoine Chambaz, Pierre Neuvial
}

