% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPVI.R
\name{getObs}
\alias{getObs}
\alias{getObs.NPVI}
\title{Retrieves the Observations}
\arguments{
\item{this}{An object of class \code{TMLE.NPVI}}

\item{tabulate}{A \code{logical}, to  specify whether it is the original 
data set that is retrieved (if \code{FALSE}) or a  tabulated version of it
(otherwise), for internal use only.  If \code{tabulate} is missing then 
the value attached to the input object is used}

\item{\dots}{Not used}
}
\value{
Either the original data set involved in the TMLE procedure or a tabulated version of it
}
\description{
Retrieves the \code{matrix} of observations involved in the TMLE procedure.
}

