% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2oEnsembleLearning.R
\name{h2o.deeplearning.Rectifier}
\alias{h2o.deeplearning.Rectifier}
\title{Deep Learning Neural Network}
\usage{
h2o.deeplearning.Rectifier(..., hidden = c(500, 500),
  activation = "Rectifier")
}
\arguments{
\item{hidden}{Parameter \code{hidden} of \code{h2o.deeplearning}, set to
\code{c(500, 500)}.}

\item{activation}{Parameter \code{activation} of \code{h2o.deeplearning},
set to "Rectifier".}

\item{\dots}{All parameters of \code{h2o.deeplearning} except \code{hidden}
and \code{activation}.}
}
\value{
Creates a H2OModel object of the right type.
}
\description{
Performs Deep Learning neural networks on an H2OFrame, with parameters
\code{activation}] and \code{hidden} set to "Rectifier" and \code{c(500,
500)}.
}
\author{
Antoine Chambaz, Pierre Neuvial
}
\seealso{
h2o.deeplearning
}

