% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2oEnsembleLearning.R
\name{h2o.deeplearning.Tanh}
\alias{h2o.deeplearning.Tanh}
\title{Deep Learning Neural Network}
\usage{
h2o.deeplearning.Tanh(..., hidden = c(200, 200, 200), activation = "Tanh")
}
\arguments{
\item{\dots}{All parameters of \code{h2o.deeplearning} except \code{hidden}
and \code{activation}.}

\item{hidden}{Parameter \code{hidden} of \code{h2o.deeplearning}, set to
\code{c(200, 200, 200)}.}

\item{activation}{Parameter \code{activation} of \code{h2o.deeplearning},
set to "Rectifier".}
}
\value{
Creates a H2OModel object of the right type.
}
\description{
Performs Deep Learning neural networks on an H2OFrame, with parameters
\code{activation}] and \code{hidden} set to "Tanh" and \code{c(200, 200,
200)}.
}
\seealso{
h2o.deeplearning
}
\author{
Antoine Chambaz, Pierre Neuvial
}
