% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2oEnsembleLearning.R
\name{h2o.glm.alpha.10}
\alias{h2o.glm.alpha.10}
\title{H2O Generalized Linear Models}
\usage{
h2o.glm.alpha.10(..., alpha = 1)
}
\arguments{
\item{\dots}{All parameters of \code{h2o.glm} except \code{alpha}.}

\item{alpha}{Parameter \code{alpha} of \code{h2o.glm}, set to \code{1.0}.}
}
\value{
Creates a H2OModel object of the right type.
}
\description{
Fit a generalized linear model, with parameter \code{alpha} set to
\code{1.0}.
}
\seealso{
h2o.glm
}
\author{
Antoine Chambaz, Pierre Neuvial
}
