% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2oEnsembleLearning.R
\name{h2o.randomForest.1000x100}
\alias{h2o.randomForest.1000x100}
\title{Builds A Random Forest Model On An H2OFrame}
\usage{
h2o.randomForest.1000x100(..., ntrees = 1000, nbins = 100)
}
\arguments{
\item{ntrees}{Parameter \code{ntrees} of \code{h2o.randomForest}, set to
\code{1000}.}

\item{nbins}{Parameter \code{nbins} of \code{h2o.randomForest}, set to
\code{100}.}

\item{\dots}{All parameters of \code{h2o.randomForest} except \code{ntrees}
and \code{nbins}.}
}
\value{
Creates a H2OModel object of the right type.
}
\description{
Builds a random forest model on an H2OFrame, with parameter \code{ntrees}
and \code{nbins} set to \code{1000} and \code{100}.
}
\author{
Antoine Chambaz, Pierre Neuvial
}
\seealso{
h2o.randomForest
}

