\name{predict.SL.DSA}
\alias{predict.SL.DSA}
\title{SL Prediction Algorithm Wrapper}
\description{Prediction algorithm wrapper for SuperLearner.}
\arguments{
  \item{object}{
}
  \item{newdata}{
}
  \item{family}{
}
  \item{X}{
}
  \item{Y}{
}
  \item{\dots}{
}
}

\value{Returns a prediction.}

\author{Antoine Chambaz, Pierre Neuvial}



\seealso{SL.DSA, SL.DSA.2}

