\name{predict.SL.glm.condExpX2givenW}
\alias{predict.SL.glm.condExpX2givenW}
\title{SL  Wrapper for Estimation of Cond. Expect. of X^2 Given W}
\description{Prediction algorithm  wrapper for SuperLearner, for the  estimation of the
conditional expectation of \eqn{X^2} given \eqn{W}.}
\arguments{
  \item{object}{
}
  \item{newdata}{
}
  \item{\dots}{
}
}

\value{Returns a prediction.}

\author{Antoine Chambaz, Pierre Neuvial}



\seealso{SL.glm.condExpX2givenW}

