\name{predict.SL.glm.condExpXYgivenW}
\alias{predict.SL.glm.condExpXYgivenW}
\title{SL  Wrapper for Estimation of Cond. Expect. of XY Given W}
\description{Prediction algorithm wrapper for SuperLearner.}
\arguments{
  \item{object}{
}
  \item{newdata}{
}
  \item{\dots}{
}
}

\value{Returns a prediction.}

\author{Antoine Chambaz, Pierre Neuvial}



\seealso{SL.glm.condExpXYgivenW}

