\name{predict.SL.glm.theta}
\alias{predict.SL.glm.theta}
\title{SL  Wrapper for Estimation of Cond. Expect. of Y Given (X,W)}
\description{Prediction algorithm wrapper for SuperLearner.}
\arguments{
  \item{object}{
}
  \item{newdata}{
}
  \item{\dots}{
}
}

\value{Returns a prediction.}

\author{Antoine Chambaz, Pierre Neuvial}



\seealso{SL.glm.theta}

