% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2oEnsembleLearning.R
\name{predict.SL.nnls}
\alias{predict.SL.nnls}
\title{h2oEnsembleLearning Wrapper For Nonnegative Least Squares Metalearning}
\usage{
\method{predict}{SL.nnls}(object, newdata, ...)
}
\arguments{
\item{object}{A fitted object as given by \code{SL.nnls}.}

\item{newdata}{The predictor variables for which predictions are wished.}

\item{\dots}{Not used.}
}
\value{
A vector of predictions for \code{newdata} derived from the
fitted object \code{object}.
}
\description{
h2oEnsembleLearning wrapper for nonnegative least squares metalearning
}
\author{
Antoine Chambaz, Pierre Neuvial
}
\seealso{
SL.nnls, method.NNLS
}

