\name{setConfLevel.NPVI}
\alias{setConfLevel.NPVI}
\alias{setConfLevel}
\title{Sets Confidence Level}
\description{Sets the confidence level of a \code{TMLE.NPVI} object.}
\usage{\method{setConfLevel}{NPVI}(this, confLevel, ...)}
\arguments{
  \item{this}{An object of class \code{TMLE.NPVI}.}
  \item{confLevel}{A \code{numeric}, confidence interval level.}
  \item{\dots}{Not used.}
}



\author{Antoine Chambaz, Pierre Neuvial}



\seealso{as.character.NPVI}
\examples{FALSE}
