% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPVI.R
\name{setConfLevel}
\alias{setConfLevel}
\alias{setConfLevel.NPVI}
\title{Sets the confidence level of a \code{TMLE.NPVI} object}
\usage{
\method{setConfLevel}{NPVI}(this, confLevel, ...)
}
\arguments{
\item{this}{An object of class \code{TMLE.NPVI}}

\item{confLevel}{A \code{numeric}, confidence interval level}

\item{...}{Not used}
}
\description{
Sets the confidence level of a \code{TMLE.NPVI} object
}
