\name{SL.DSA.2}
\alias{SL.DSA.2}
\title{SL Prediction Algorithm Wrapper}
\description{Prediction algorithm wrapper for SuperLearner.}
\arguments{
  \item{\dots}{
}
  \item{X}{
}
  \item{maxsize}{
}
  \item{maxorderint}{
}
  \item{maxsumofpow}{
}
  \item{Smove}{
}
  \item{vfold}{
}
}

\value{Returns a fitted object.}

\author{Antoine Chambaz, Pierre Neuvial}



\seealso{SL.DSA, predict.SL.DSA}

