\name{SL.DSA}
\alias{SL.DSA}
\title{SL Prediction Algorithm Wrapper}
\description{Prediction algorithm wrapper for SuperLearner.}
\arguments{
  \item{Y}{
}
  \item{X}{
}
  \item{newX}{
}
  \item{family}{
}
  \item{obsWeights}{
}
  \item{maxsize}{
}
  \item{maxorderint}{
}
  \item{maxsumofpow}{
}
  \item{Dmove}{
}
  \item{Smove}{
}
  \item{vfold}{
}
  \item{\dots}{
}
}

\value{Returns a fitted object.}

\author{Antoine Chambaz, Pierre Neuvial}



\seealso{SL.DSA.2, predict.SL.DSA}

