\name{SL.glm.condExpX2givenW}
\alias{SL.glm.condExpX2givenW}
\title{SL  Wrapper for Estimation of Cond. Expect. of X^2 Given W}
\description{Prediction algorithm  wrapper for SuperLearner, for the  estimation of the
conditional expectation of \eqn{X^2} given \eqn{W}.}
\arguments{
  \item{Y}{
}
  \item{X}{
}
  \item{newX}{
}
  \item{family}{
}
  \item{obsWeights}{
}
  \item{\dots}{
}
}

\value{Returns a fitted object.}

\author{Antoine Chambaz, Pierre Neuvial}



\seealso{learnCondExpX2givenW, predict.SL.glm.condExpX2givenW}

