\name{SL.glm.condExpXYgivenW}
\alias{SL.glm.condExpXYgivenW}
\title{SL  Wrapper for Estimation of Cond. Expect. of XY Given W}
\description{Prediction algorithm  wrapper for SuperLearner, for the  estimation of the
conditional expectation of \eqn{XY} given \eqn{W}.}
\arguments{
  \item{Y}{
}
  \item{X}{
}
  \item{newX}{
}
  \item{family}{
}
  \item{obsWeights}{
}
  \item{\dots}{
}
}

\value{Returns a fitted object.}

\author{Antoine Chambaz, Pierre Neuvial}



\seealso{learnCondExpXYgivenW, predict.SL.glm.condExpXYgivenW}

