\name{SL.glm.g}
\alias{SL.glm.g}
\title{SL  Wrapper for Estimation of Cond. Prob. of X=0 Given W}
\description{Prediction algorithm  wrapper for SuperLearner, for the  estimation of the
conditional probability of \eqn{X=0} given \eqn{W}.}
\arguments{
  \item{Y}{
}
  \item{X}{
}
  \item{newX}{
}
  \item{family}{
}
  \item{obsWeights}{
}
  \item{\dots}{
}
}

\value{Returns a fitted object.}

\author{Antoine Chambaz, Pierre Neuvial}



\seealso{learnG, predict.SL.glm.g}

