\name{SL.glm.theta}
\alias{SL.glm.theta}
\title{SL  Wrapper for Estimation of Cond. Prob. of X=0 Given W}
\description{Prediction algorithm  wrapper for SuperLearner, for the  estimation of the
conditional expectation of \eqn{Y} given \eqn{(X,W)}.}
\arguments{
  \item{Y}{
}
  \item{X}{
}
  \item{newX}{
}
  \item{family}{
}
  \item{obsWeights}{
}
  \item{\dots}{
}
}

\value{Returns a fitted object.}

\author{Antoine Chambaz, Pierre Neuvial}



\seealso{learnTheta, predict.SL.glm.theta}

