% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperLearning.R
\name{predict.SL.glm.condExpX2givenW}
\alias{predict.SL.glm.condExpX2givenW}
\title{SL Wrapper for Estimation of Cond. Expect. of X^2 Given W}
\usage{
\method{predict}{SL.glm.condExpX2givenW}(object, newdata, ...)
}
\arguments{
\item{object}{A fitted object as given by \code{SL.glm.condExpX2givenW}.}

\item{newdata}{The predictor variables for which predictions are wished.}

\item{\dots}{Not used.}
}
\value{
A vector of predictions for \code{newdata} derived from the fitted
object \code{object}.
}
\description{
Prediction algorithm wrapper for SuperLearner, for the estimation of the
conditional expectation of \eqn{X^2} given \eqn{W}.
}

