% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperLearning.R
\name{predict.SL.glm.g}
\alias{predict.SL.glm.g}
\title{SL Wrapper for Estimation of Cond. Prob. of X=0 Given W}
\usage{
\method{predict}{SL.glm.g}(object, newdata, ...)
}
\arguments{
\item{object}{A fitted object as given by \code{SL.glm.g}.}

\item{newdata}{The predictor variables for which predictions are wished.}

\item{\dots}{Not used.}
}
\value{
A vector of predictions for \code{newdata} derived from the fitted
object \code{object}.
}
\description{
Prediction algorithm wrapper for SuperLearner, for the estimation of the
conditional probability of X=0 given W
}

